backend/requirements.txt
Django
chromadb
mysqlclient

djangorestframework
djangorestframework-simplejwt